Add /blog back to project
