+++ eslint.config.js
@@ -1,11 +1,9 @@
 import js from '@eslint/js';
-import globals from 'globals';
 
 export default [
   // Use ESLint recommended
   js.configs.recommended,
   {
-    languageOptions: { globals: { ...globals.node } },
     rules: {
       'max-nested-callbacks': ['error', 3],
       'no-return-await': 'error',
