+++ package.json
@@ -1,7 +1,6 @@
 {
   "name": "jrisner",
   "type": "module",
-  "version": "1.0.0",
   "license": "UNLICENSED",
   "scripts": {
     "format": "prettier . --check",
@@ -16,9 +15,9 @@
     "@astrojs/sitemap": "^3.2.1",
     "@astrojs/tailwind": "^5.1.3",
     "@playwright/test": "^1.49.1",
-    "astro": "^4.16.16",
+    "astro": "^5.0.8",
     "autoprefixer": "^10.4.20",
-    "eslint": "^9.17.0",
+    "eslint": "^9.16.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.2.1",
     "markdownlint-cli": "^0.43.0",
