+++ src/components/navigation/Header.astro
@@ -14,6 +14,9 @@ import Logo from './Logo.astro';
           <a href="/" class="hover:text-muted" data-testid="navLink">Home</a>
         </li>
         <li>
+          <a href="/blog" class="hover:text-muted" data-testid="navLink">Blog</a>
+        </li>
+        <li>
           <a href="/cheat-sheets" class="hover:text-muted" data-testid="navLink">Cheat Sheets</a>
         </li>
       </ul>
