+++ src/content/blog/python-development-environment.md
@@ -1,0 +1,8 @@
+---
+description: 'What I use when developing in Python.'
+heroImage: '/images/starry-night.webp'
+isDraft: false
+publishDate: '2024-12-16'
+tags: ['python']
+title: 'Python Development Environment'
+---
