+++ src/pages/blog/[...slug].astro
@@ -1,0 +1,33 @@
+---
+import { getCollection, render } from 'astro:content';
+import BaseLayout from '@layouts/BaseLayout.astro';
+import { SITE_TITLE } from 'src/config';
+
+export const getStaticPaths = async () => {
+  const publishedPosts = await getCollection('blog', ({ data }) => !data.isDraft);
+
+  return publishedPosts.map((post) => ({
+    params: { slug: post.id },
+    props: { post }
+  }));
+};
+
+const { post } = Astro.props;
+const { title, publishDate, heroImage } = post.data;
+const { Content } = await render(post);
+---
+
+<BaseLayout title={`${title} | ${SITE_TITLE}`}>
+  <section class="mx-auto mb-12 max-w-blog text-center" data-testid="postSummary">
+    <h1 class="text-3xl font-bold md:text-5xl">{title}</h1>
+    <div class="mb-8 mt-2">
+      <time class="text-sm text-muted md:text-base" datetime={publishDate.toISOString()}>
+        {publishDate.toLocaleDateString('en-us', { day: '2-digit', month: 'long', year: 'numeric', timeZone: 'UTC' })}
+      </time>
+    </div>
+    <img src={heroImage} alt="Post Hero Image" class="rounded-2xl" />
+  </section>
+  <article id="markdown-content" class="max-w-blog">
+    <Content />
+  </article>
+</BaseLayout>
