+++ src/pages/blog/index.astro
@@ -1,0 +1,31 @@
+---
+import { getCollection } from 'astro:content';
+import BaseLayout from '@layouts/BaseLayout.astro';
+import BlogPostCard from '@components/blog/BlogPostCard.astro';
+import { SITE_TITLE } from 'src/config';
+
+const publishPosts = await getCollection('blog', ({ data }) => !data.isDraft);
+const posts = publishPosts.sort((a, b) => b.data.publishDate.valueOf() - a.data.publishDate.valueOf());
+---
+
+<BaseLayout title={`Blog | ${SITE_TITLE}`}>
+  <h1 class="text-3xl font-bold md:text-5xl">Posts</h1>
+  <section>
+    <ul class="divide-y divide-slate-700">
+      {
+        posts.map(({ data, id: slug }) => (
+          <li class="py-8 md:py-12">
+            <BlogPostCard
+              title={data.title}
+              description={data.description}
+              image={data.heroImage}
+              publishDate={data.publishDate}
+              tags={data.tags}
+              target={`/blog/${slug}`}
+            />
+          </li>
+        ))
+      }
+    </ul>
+  </section>
+</BaseLayout>
