+++ src/pages/cheat-sheets/[...slug].astro
@@ -1,20 +1,20 @@
 ---
 import BaseLayout from '@layouts/BaseLayout.astro';
-import { getCollection } from 'astro:content';
+import { getCollection, render } from 'astro:content';
 import { SITE_TITLE } from 'src/config';
 
 export const getStaticPaths = async () => {
-  const publishedSheets = await getCollection('cheat-sheets', ({ data }) => data.isDraft === false);
+  const publishedSheets = await getCollection('cheatSheets', ({ data }) => !data.isDraft);
 
   return publishedSheets.map((sheet) => ({
-    params: { slug: sheet.slug },
+    params: { slug: sheet.id },
     props: { sheet }
   }));
 };
 
 const { sheet } = Astro.props;
 const { title } = sheet.data;
-const { Content } = await sheet.render();
+const { Content } = await render(sheet);
 ---
 
 <BaseLayout title={`${title} | ${SITE_TITLE}`}>
