+++ test/navigation.spec.ts
@@ -1,8 +1,9 @@
 import { test, expect } from '@playwright/test';
 
+const expectedLinks = ['Home', 'Blog', 'Cheat Sheets'];
+
 test.describe('Navigation - Large Viewport', () => {
   test('navigates to each page successfully', async ({ page }) => {
-    const expectedLinks = ['Home', 'Cheat Sheets'];
     const allNavLinks = page.getByTestId('navLink');
 
     await page.goto('/');
@@ -26,7 +27,6 @@ test.describe('Navigation - Large Viewport', () => {
 
 test.describe('Navigation - Mobile Viewport', () => {
   test('displays navigation overlay with nav links', async ({ page }) => {
-    const expectedLinks = ['Home', 'Cheat Sheets'];
     const navToggle = page.getByTestId('mobileNavToggle');
 
     await page.setViewportSize({ width: 767, height: 940 });
